internal/storage: make lock test deterministic
Rewrite the test that verifies that DB.Lock and DB.Unlock behave
correctly, to make it deterministic and independent of the race detector.

Also, put it in a separate function. Despite being deterministic,
it is still a problem for record/replayers, because of the delay
it requires to confirm that the lock is held (on
internal/storage/test.go:122). In the actual program, the db.Lock in the
goroutine on line 112 is blocked by the db.Unlock on line 125. But a
record/replayer doesn't know about causality, so the program will keep
running, attempting to make RPCs that may not be present in the trace.

Adding timing delays to a replayer in an attempt to mimic the behavior
of the original program will help, but there will still be a race
between the Lock in the goroutine proceeding, and the Unlock.

Change-Id: Ia27a228e09be39e0756b60cda2abcaf38dc9a327
Reviewed-on: https://go-review.googlesource.com/c/oscar/+/599695
Reviewed-by: Zvonimir Pavlinovic <zpavlinovic@google.com>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Reviewed-by: Russ Cox <rsc@golang.org>
Reviewed-by: Tatiana Bradley <tatianabradley@google.com>