internal/gcp/firestore: improve vector DB test
Restore the function argument to TestVectorDB to return
a reference to the same database, as intended.

Comment storage.TestVectorDB to clarify that intent.

Delete vectors before recording.

Also, upgrade to a new grpcreplay version that will reduce
spurious diffs in the future (but not in this CL) by
removing excess spaces from the recording.

Change-Id: I113e2450a0b45b1ee23a528a00547b3b40fedfaa
Reviewed-on: https://go-review.googlesource.com/c/oscar/+/603636
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Reviewed-by: Zvonimir Pavlinovic <zpavlinovic@google.com>