internal/gcp/gcphandler: more information in message
The GCP log viewer displays only the message of a log entry on the screen.
To see the attributes requires two clicks, one to open the entry and
a second to open its "jsonPayload" field.

To mitigate the problem, this handler appends the initial attributes
to the message, up to a fixed length.

Change-Id: I9aae87ca589f8e66751e9827c1514b2c37db30c0
Reviewed-on: https://go-review.googlesource.com/c/oscar/+/607136
Reviewed-by: Tatiana Bradley <tatianabradley@google.com>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>