internal/gcp: copy gemini and grpcrr packages
This is a partial move of gemini and grpcrr to gcp. grpcrr is completely
moved: it is copied to gcp and its original is deleted. The latter is
made possible by the fact that only gcp packages use grpcrr.

In a follow-up CL, the original gemini package will be delete and gaby
module will use the gemini from gcp.

Updates golang/oscar#5

Change-Id: Ied502ce000a9ea5f6d8acc36e2692550c3fa19f5
Reviewed-on: https://go-review.googlesource.com/c/oscar/+/606875
Reviewed-by: Jonathan Amsterdam <jba@google.com>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>