internal: add metric for Watchers
Export a metric for the latest time of gaby Watchers.
This will let us track their progress.

A single metric is used for all Watchers, with the different
watchers distinguished by an attribute called "name".
Having one metric is convenient, but the default GCP "Sum" aggregation
is a poor choice; it is better to view the metrics unaggregated.

The Watcher.cutoff method returns the latest DBTime, but it requires
the Watcher lock to be held, and it may be held for a while. It's
not a good idea to wait a long time for a lock during metric collection,
so instead we track the latest time in an atomic variable and expose
its value with the Watcher.Latest method. The various users of Watchers
themselves provide Latest functions or methods to their users.
Ultimately the gaby main program is responsible for collecting
and naming all the Watcher values and registering the metric.

This CL also refactors a bit, removing the convenience layer that was
hiding Open Telemetry from the gcpmetrics package and giving that job to the
gaby program.

Change-Id: Ic69e647a6f259e0a5b38bdc246992e9078a64fad
Reviewed-on: https://go-review.googlesource.com/c/oscar/+/607155
Reviewed-by: Tatiana Bradley <tatianabradley@google.com>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>