internal/{gaby,github}: add gaby endpoint github-event (webhook listener)
Add a gaby endpoint, /github-event, which listens for
POST requests from GitHub webhooks and logs the payload
if the request is valid.

(This can be extended to take immediate action when,
for example, a new issue is posted.)

The validation logic is in the internal/github package.
The request is validated by checking the payload against
the HMAC in the request header, using a secret stored
in GCP secret manager.

Change-Id: I96f6779799be702ee8dc1b7631cce5bd3eaad3f1
Reviewed-on: https://go-review.googlesource.com/c/oscar/+/610335
Reviewed-by: Jonathan Amsterdam <jba@google.com>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>