internal/gaby: reorganize the search code
Have doSearch return the contents of the page, instead of writing it.

The difference is small, but eliminates the possibility that doSearch
will write to the ResponseWriter and then return an error. If that
were to happen, the user would see a 200 status, not a 500, because
the header is sent just before the first byte is written.

Change-Id: I0f9686b337526526f1532bb74e52f58dfac87ed7
Reviewed-on: https://go-review.googlesource.com/c/oscar/+/612095
Reviewed-by: Tatiana Bradley <tatianabradley@google.com>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>