all: propagate important errors in Runs and Syncs
Run and Sync actions currently only log errors, regardless of their
importance. This is likely due to the fact that, until now, there was no
place to surface important errors rather than the logs or crashing the
whole application. This is not the case as of recently, when we added
support for cloud Error Reporting. This CL hence propagates and tests
important Run and Sync errors and surfaces them to Error Reporting
service in cron endpoint.

Updates golang/oscar#12

Change-Id: I1700e74854f42f72363080484a5a3f22fc7e675e
Reviewed-on: https://go-review.googlesource.com/c/oscar/+/612375
Reviewed-by: Jonathan Amsterdam <jba@google.com>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>