internal/actions: add After
Originally, Before returned a function that could be called
after the action. But that won't work for actions that are approved
and executed later, perhaps in another process.

Now Before returns a key that can be passed to After. The same
key can be obtained from an Entry, which is what you'll get if
you scan for approved actions. So After works in both cases.

I toyed with a few other designs before picking this one. The orginal
one in the design had Before returning one part of the key, requiring
code to keep track of the other two. I also considered having Before
return an Entry which After would take as an argument, but that
might give the impression that changes to the Entry's fields would
be saved, which is wrong and would be a bad idea.

Change-Id: I5d7b9e224f5a5659fc651682fc569dbf909a76fe
Reviewed-on: https://go-review.googlesource.com/c/oscar/+/612439
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Reviewed-by: Tatiana Bradley <tatianabradley@google.com>