internal/actions: rename namespace to actionKind
Renamed "namespace" to "actionKind".

The word "namespace" implies that there are separate named "spaces"
each with the same structure. What we were calling a namespace is really
a short description of the kind of action that is represented, which
determines the key and the Action and Result fields.

Change-Id: I0355b33c70b06df344b9998057f98136ee37f280
Reviewed-on: https://go-review.googlesource.com/c/oscar/+/612440
Reviewed-by: Tatiana Bradley <tatianabradley@google.com>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>