internal/actions: support wall clock scanning
It's common to want to look at the last hour of logs,
or the logs between two times. The times in those queries are
expressed as everyday ("wall clock") times, but the timed
package uses the DBTime type, which need not be related
to the wall clock time (although currently it is).

To support those kinds of queries, this CL creates a relationship
between wall clock and DBTimes. Each time an entry is added
to the DB, a new association is made between the entry's created
or done time and the DBTime under which it is stored by the timed
package.

The ScanAfter function scans by wall clock time, by using the
mapping to obtain the corresponding DBTime and scanning with that.
The old ScanAfter is renamed ScanAfterDBTime, anticipating that scanning
that way will be less common.

Change-Id: Ib4eaa5004fc9424575962b3832d5ce175b50e552
Reviewed-on: https://go-review.googlesource.com/c/oscar/+/612675
Reviewed-by: Tatiana Bradley <tatianabradley@google.com>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>