internal/related: add related.Post to post to single issue
Add a function [related.Post] which posts related issues
for a single issue. It does not monitor or modify the watcher
used by [related.Run].

This will be used to respond quickly to individual events,
for example in response to a GitHub webhook event.

This CL refactors the code for [related.Post] in order
to re-use its logic.

Change-Id: I470652e7b3a3ee173479369e72c07b7c53ffe986
Reviewed-on: https://go-review.googlesource.com/c/oscar/+/612515
Reviewed-by: Jonathan Amsterdam <jba@google.com>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>