internal/gaby: refactor gaby cron function and add locks
Refactor the function run by the /cron endpoint
to make it an explicit method of Gaby instead
of creating it inside the main() function. Store
the values needed to run the method (related.Poster
and commentfix.Fixer) as fields of Gaby.

This makes it more clear what order the actions
should be performed in (which sometimes matters:
for example, embedding should happen last), and
additionally makes it easier for future endpoints
of Gaby to specify which individual actions to run.

Also add locks for logical groups of events to
avoid a race between /cron and /githubevent.
The endpoints can run concurrently, so we need
to make sure they aren't accessing the same watchers
at the same time.

The locks are at the granularity of
independent syncs/actions so we can easily
run single actions in the future.

For example, the GitHub webhook will only handle the actions
directly related to an issue event, so it needs to
pick a subset of actions to perform. Also, it shouldn't
need to wait for /cron to finish completely before
starting.

Change-Id: Ibb990bcdeb5cd9597a59217a2205f8c5f5487d50
Reviewed-on: https://go-review.googlesource.com/c/oscar/+/612516
Commit-Queue: Tatiana Bradley <tatianabradley@google.com>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Reviewed-by: Zvonimir Pavlinovic <zpavlinovic@google.com>