internal/gcp/firestore: fix 'lock of held lock' bug
Fix a bug in which concurrent locks of the same lock
could sometimes panic instead of waiting.

This was caused because the "held" bool in a
firestore transactions was not reset explicitly
in each run. Because firestore transactions
can run multiple times, sometimes "held" was
incorrectly set to true by a stale run.

Change-Id: Id0a5b1b91e862d57fda8d10d7b8e94dd7c7a1abf
Reviewed-on: https://go-review.googlesource.com/c/oscar/+/613176
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Reviewed-by: Zvonimir Pavlinovic <zpavlinovic@google.com>