gaby: add /crawl endpoint
Add a new Gaby endpoint, /crawl, which runs the crawler.

Currently, this can only be triggered manually. Once we know
how fast it is, we can set up a Cloud scheduler job that
triggers it at an appropriate cadence.

It is decoupled from the /cron endpoint as we expect it
to be slow, and it is not necessary for all crawled webpages
to be up-to-date in the docs corpus to take actions.

For golang/oscar#26

Change-Id: I14b9203c79b188a4a0d728844770556380a6f858
Reviewed-on: https://go-review.googlesource.com/c/oscar/+/614162
Reviewed-by: Jonathan Amsterdam <jba@google.com>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>