internal/{gaby,github}: move webhook project check to gaby
Move the check for the expected project of a webhook event
from the github package to the gaby package.

This allows the handleGitHubEvent function to log a
warning instead of an error if the project isn't correct.

(This should only happen if an authorized user mistakenly
set up a webhook for the wrong project, or forgot to suspend
a webhook for an old project. A non-authorized user would
not know the keys required to set up a webhook for an arbitrary
project.)

Change-Id: I68b92527109fabafcf970df63f51fbd0d0c2e791
Reviewed-on: https://go-review.googlesource.com/c/oscar/+/614637
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Reviewed-by: Jonathan Amsterdam <jba@google.com>