internal/{gaby,github}: no error for unexpected GitHub event
Log an info message and ignore the event instead of
returning an error in [Gaby.handleGitHubEvent] if a
webhook event is of an unknown type. (Receiving a
notification for an event we don't care about should not
be a hard error).

To achieve this, modify the behavior of github.ValidateWebhookRequest
to store a json.RawMessage as the payload of events of
unknown type, instead of returning an error.

Change-Id: I7b22d2f7d35269d581dc28496e4daad2e0422274
Reviewed-on: https://go-review.googlesource.com/c/oscar/+/614755
Reviewed-by: Jonathan Amsterdam <jba@google.com>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>