remove unique part of action key
The actions log was appending a random number to each key to
make sure they were unique.

The intention was to make sure that every attempt to log an action
generated a new log entry. But if we don't do that and let the user
control the key, then the action log can double as way to check if
a piece of work has been done already, and avoid doing it again.

For example, the comment fixer records whether it has fixed an issue
or comment in the DB, to avoid duplicating work. Prior to this change,
it would need to continue doing that as well as logging actions.
But as a result of this change and the upcoming CL, it won't need those
extra DB entries, because the action log will record whether the action
was done.

(The next CL will make this clearer.)

For #9.

Change-Id: Iede5df4184e6bffaa75ac7d2fdb4ece4b1fb8f0f
Reviewed-on: https://go-review.googlesource.com/c/oscar/+/615615
Reviewed-by: Tatiana Bradley <tatianabradley@google.com>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Reviewed-by: Zvonimir Pavlinovic <zpavlinovic@google.com>