internal/actions: Register and Run
Replace Before and After with Register and Run.

Register is called by components that want to use the log,
and Run should be called by the main program to run all actions.

See the top comment in actions.go for more.

For golang/oscar#9.

For #9.

Change-Id: Id8dea631680b28ebae98414325741453a3615dfe
Reviewed-on: https://go-review.googlesource.com/c/oscar/+/615616
Reviewed-by: Tatiana Bradley <tatianabradley@google.com>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>