internal/gerrit: make partial progress backwards
Gerrit APIs list changes only in reverse chronological order. Since the
execution of the sync can be interrupted at any point in time, there
needs to be a mechanism for partial progress while going backwards. This
changes implements that mechanism.

The idea is to go backwards and keep track of how far we've analyzed in
the past, where we started from, and the place where we are in between
(in case there is an interruption). Once there is nothing to analyze in
the past, the algorithm sets the start position as the new past value
and takes the current moment in time as the new start position.

Change-Id: Ie56e05748a5c62d3a4426df92e3806e6a6af4d49
Reviewed-on: https://go-review.googlesource.com/c/oscar/+/615680
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Reviewed-by: Jonathan Amsterdam <jba@google.com>
Reviewed-by: Ian Lance Taylor <iant@google.com>