internal/actions, etc.: registrants control display
Improve the display of actions on the actionlog page, and with any
future tools.

Instead of displaying the action as its original JSON, let the
component that registers with the log control the action's format.

Since there are now two functions to register instead of just one,
use an interface.

The main display change in this CL is the use of a diff for comment
fixers, rendering the changes clearer.

For golang/oscar#9.

Change-Id: I39bf91aac741595f3833f5b132ddfdc13c25672b
Reviewed-on: https://go-review.googlesource.com/c/oscar/+/617436
Reviewed-by: Zvonimir Pavlinovic <zpavlinovic@google.com>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>