internal/reviews: new package
This is a step toward a review dashboard.

The reviews package includes a view of changes that is independent
of a particular revision control system. It also includes an
implementation of that view for Gerrit. In practice any specific
Gerrit project, such as the Go project, will need to enhance that
implementation with information specific to that project.

Change-Id: I9aae301a1c52cf1fc77102d0d5b83e38cdfde0c1
Reviewed-on: https://go-review.googlesource.com/c/oscar/+/617760
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Auto-Submit: Ian Lance Taylor <iant@golang.org>
Reviewed-by: Ian Lance Taylor <iant@google.com>
Reviewed-by: Jonathan Amsterdam <jba@google.com>