internal/{content,adapters}: initial sketch of project model
Add types that generalize concepts like issue, issue comment,
discussion and so on. Consider them all to be pieces of content.
A subtype of Content is Post, an easily posted and edited
item like a GitHub issue or comment.

Begin work on a GitHub Adapter, which will provide all access
to GitHub and the synced GitHub data in the database.

Provide a GitHub issues source, which abstracts CRUD operations
on GitHub itself.

The change to internal/github/edit.go adds a return value to
PostIssueComment. The new return value is the issue URL for the API,
in addition to the display URL (for people). We want to use the
API URL as the unique identifier for the issue comment.

Change-Id: Ic0ae8db92049a1a41bd441e1f610b67340754dd7
Reviewed-on: https://go-review.googlesource.com/c/oscar/+/621695
Reviewed-by: Zvonimir Pavlinovic <zpavlinovic@google.com>
Reviewed-by: Hyang-Ah Hana Kim <hyangah@gmail.com>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Reviewed-by: Tatiana Bradley <tatianabradley@google.com>