internal/httpr: add scrubbing functions for responses
The obvious approach is to have scrubbing functions work on
http.Responses directly. The problem is that the original response has
to be kept, so scrubbing can work on only on clones. Clones can have
.Close set to true [1] even when the original response had it false.
Setting that field explicitly to false does not work, since Write method
on responses reintroduces it internally [2]. We need to invoke the Write
method to save the string representation of a response for replaying.

The solution is for the scrubbing to work on byte stream of the response
and document possible cloning issues.

[1] go/src/net/http.readTransfer:557
[2] go/src/net/http.Response.Write:292

Change-Id: Iaaababdef52a004b7c9be7ff4d328ef0addad083
Reviewed-on: https://go-review.googlesource.com/c/oscar/+/621916
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Reviewed-by: Jonathan Amsterdam <jba@google.com>