internal/llmapp: add cache for LLM responses
Add a mechanism to cache LLM responses. This will make the
Overviews page more usable when used by multiple people at the
same time (e.g., during proposal review).

Cached responses are based on the generative model and prompt, so
in particular any changes to the prompt (such as a new document) would
be a cache miss.

To support this, refactor the llmapp package to expose a Client
on which all the Overview methods are defined.

Change-Id: Ie688613437f70dde116630e85b1041c2c5709bfd
Reviewed-on: https://go-review.googlesource.com/c/oscar/+/623355
Reviewed-by: Zvonimir Pavlinovic <zpavlinovic@google.com>
Reviewed-by: Jonathan Amsterdam <jba@google.com>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>