internal/github: fix parsing of times
The format for parsing times differed from the one used by the GitHub
API. The error was silently lost. Instead, a zero time was returned,
meaning all issues and comments appeared to be before the
commentfix.Fixer's cutoff and were not processed.

This CL should fix the problem for subsequent issues, but we still
have to go back and manually fix the ones that were missed somehow.

Besides using the right format, and testing it against actual values
from GH, also panic on error so gaby fails fast. It was only because
Cherry was observant that this was noticed at all.

For golang/oscar#54.

Change-Id: Ib639dd68cae950ccba1af8e798ea6ae2c9ba5e19
Reviewed-on: https://go-review.googlesource.com/c/oscar/+/629797
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Reviewed-by: Hyang-Ah Hana Kim <hyangah@gmail.com>