internal/gaby: quick fix for markdown rendering error
Add a fix for a specific markdown error (made by the LLM) found in
the overview of issue 70471, in which a citation was placed on the
same line as the backticks indicating the end of a code block.

If we notice numerous markdown errors in the future, we should
come up with a more robust solution.

(To aid in future debugging, this CL adds a UI element that allows
the user to view the raw, unfixed output of the LLM).

Change-Id: I7d225e0304d97c75cd5f3de7c6f7b513a21d070d
Reviewed-on: https://go-review.googlesource.com/c/oscar/+/630697
Reviewed-by: Hyang-Ah Hana Kim <hyangah@gmail.com>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>