internal/{gaby,github,llmapp}: add feature to summarize updates to a discussion
Add a new feature to the Overviews page to summarize the new comments
posted to a thread since a certain comment.

The goal is to allow someone (e.g., a proposal reviewer) to get up to
speed with new comments they haven't seen yet.

For now, the user needs to specify the comment ID of the last comment
they have read. We probably want to make this more user-friendly
in the future.

There is some small refactoring to support this.

Change-Id: I165604e5fc4702584fc493ebdc7a240da500d59f
Reviewed-on: https://go-review.googlesource.com/c/oscar/+/631197
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Reviewed-by: Jonathan Amsterdam <jba@google.com>