internal: generate structured output for related docs analysis
Specify a JSON structure for the LLM to follow when generating
the related docs overview. This will allow us to choose which
order the related documents appear in, what headings are used, etc.
(In particular, we can display the documents in the same order
and with the same headings as gabyhelp's related post).

This change is mostly refactoring, but also adds a small new task
for the LLM, which is to guess the relevance of a related doc to an
original doc. We may use this in the future to trim down or re-order
the documents that Gaby posts.

Refactor to support this change.

Change-Id: If64739bb4fecb3cb0b292fbaa647972cf57299b6
Reviewed-on: https://go-review.googlesource.com/c/oscar/+/631776
Reviewed-by: Jonathan Amsterdam <jba@google.com>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>