internal/codeimage: extract code from images
The codeimage package attempts to extract code from images using an
LLM.

The LLM's output is checked for syntactic validity, but it may not
compile. We can't actually compile it without a sandbox.

Perhaps in a followup we can typecheck it with go/types.
I'm not sure if we can use go/packages because it may invoke the compiler
for some configurations.

For #67.


Change-Id: Ibb9b65a6bdf7fc2cefa0a6eb341939b4a7456c89
Reviewed-on: https://go-review.googlesource.com/c/oscar/+/634295
Reviewed-by: Zvonimir Pavlinovic <zpavlinovic@google.com>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Reviewed-by: Hyang-Ah Hana Kim <hyangah@gmail.com>