internal/github: add Wrapper for github edits
A Wrapper is used to wrap comments/edits made to GitHub with hidden
tags so that they can later be identified without relying on a database.

The first use of this will be by the Overview poster to wrap its own
comments and edits, so it can easily identify and strip out its own edits
from a body of text.

Change-Id: I8aa1f14d0d68264be4221f5fa99b5b24278d153f
Reviewed-on: https://go-review.googlesource.com/c/oscar/+/636858
Reviewed-by: Hyang-Ah Hana Kim <hyangah@gmail.com>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>