internal/labels: save assigned categories in DB
When the labeler assigns labels to an issue, store the category
names in the DB.

We store the categories instead of the labels because we control them,
so we can ensure that they don't change over time or across projects.

For golang/oscar#64.

Change-Id: I857d5ebbc0c761c0c59b4a3ab7ef203278bd8384
Reviewed-on: https://go-review.googlesource.com/c/oscar/+/639015
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Reviewed-by: Ian Lance Taylor <iant@google.com>