internal/{gcp/checks,llm}: make LLM policies fixed instead of configurable
Modify the [llm.PolicyChecker] interface to require that the policies
configured on a checker are fixed throughout its lifetime. Modify the
implementations accordingly.

The purpose of this change is to make it easier to cache the results
of a policy checker (in a follow up CL) without needing to prevent
race conditions related to changing policies. There's no need in the
current code to modify the policies on the fly.

Change-Id: I475113f05152d8d37c8036ca8cd4359f8081714f
Reviewed-on: https://go-review.googlesource.com/c/oscar/+/640835
Reviewed-by: Jonathan Amsterdam <jba@google.com>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>