internal/llmapp: update formatting of overviews to look better on GitHub
Update the formatting of overview headings to display better when
posted as comments on GitHub. This would be more robust to modify
in the future if we used a JSON schema instead of an unstructured prompt,
so leave that as a TODO for future work.

Change-Id: I21e50b673090ab5b09e66f972d785fb700c078fc
Reviewed-on: https://go-review.googlesource.com/c/oscar/+/641175
Reviewed-by: Jonathan Amsterdam <jba@google.com>
Auto-Submit: Tatiana Bradley <tatianabradley@google.com>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>