internal/labels: add examples to prompt
Add examples of issues and their correct categories to the prompt
of the LLM request, to improve accuracy.

Also change the args to labeleval to take a project instead of a
category file. Adding examples to labeleval proved too unwieldy.
A user can still experiment with different categories and examples
simply by editing the files under internal/labels/static in a branch.

After turning some recent failures into examples, the score on the
labeleval corpus increased from 68.9% to 84.4%. Admittedly, that is
cheating, since the examples appear in the corpus! But it does
demonstrate that the examples are working.

Interestingly, the LLM decided that https://go.dev/issue/71132 is
an implementation request, despite that being an example labeled
featureRequest. I have to say, it's not wrong; that could go either way.
Leaving it for now.

Change-Id: Ifa1cb3a3a26d198616edeb2fa6d24ee84a3fb19d
Reviewed-on: https://go-review.googlesource.com/c/oscar/+/641695
Reviewed-by: Hyang-Ah Hana Kim <hyangah@gmail.com>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>