internal/sandbox: create unique sandbox for command invocation
There is no guarantee that a sandboxed command will execute in a
non-concurrent setting. Hence, we need to create a unique sandbox for
each command and each of its invocations. We will often use the same
command (e.g., "goimports") but in a different directory. We create a
unique sandbox based on the command itself, its working directory, and
current time for sanity.

Change-Id: Ic02f64b4eda2d34f0eb80a81c971dd0ee0143db7
Reviewed-on: https://go-review.googlesource.com/c/oscar/+/641535
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Reviewed-by: Jonathan Amsterdam <jba@google.com>
Reviewed-by: Andres felipe Becerra <andbecc1991@gmail.com>