internal/gaby/Dockerfile: autoapprove labels
No longer require manual approval for labels.

Based on recent issues, the labeler is doing a decent job.
It is not perfect, but it will never be perfect.
Also, I can't keep up with Sean and other triagers. The labeler isn't
useful unless it labels issues before a human does so.
And the best way to get feedback on its performance is to measure when
people remove or change its choices.

Change-Id: Id9495b438d3ce71e0af800b0278402054c401ee1
Reviewed-on: https://go-review.googlesource.com/c/oscar/+/642080
Reviewed-by: Hyang-Ah Hana Kim <hyangah@gmail.com>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Reviewed-by: Tatiana Bradley <tatianabradley@google.com>
Reviewed-by: Zvonimir Pavlinovic <zpavlinovic@google.com>